goepay: verify checksum using hmac_sha1 in hex
The checksum is now performed using hmac_sha1 of the requested values
and comparison of the result with the provided CHECKSUM query parameter.

If checksum is not matching then bad request is returned with indication
that checksum is not available. The check or payment processing is
performed only when checksum is matching.