goepay: update code to be compatible with go 1.13
The existing code is now updated to be compatible with go 1.13 and could
be deployed on latest GAE.

As the old code will be deprecated in the next few months, the old
structure is completely removed.

As side change the go modules are now used to provide better dependency
management.